+++ src/ModifiersWorker.ts
@@ -2,13 +2,18 @@ import { getAllMethodsFromModule } from './helpers/module'
 import { StonexModule } from './StonexModule'
 import { Store } from './StonexStore'
 
-export declare type StoreModifier<MP, D = any> = (store: Store<MP> | null) => (void | D)
-export declare type ModuleModifier<D = any> = (module: StonexModule) => (void | D)
+export declare type StoreModifier<MP, D = ModuleModifier> = (store: Store<MP> | null) => (void | D)
+export declare type ModuleModifier<D = ActionModifier> = (module: StonexModule) => (void | D)
 export declare type ActionModifier = (args: any[], moduleName: string, methodName: string) => false | any
 
-export declare type Modifier<MP> = StoreModifier<MP, ModuleModifier<ActionModifier>>
+export declare type Modifier<MP> = StoreModifier<MP>
 
-export default class ModifiersWorker {
+/**
+ * Internal class which usings only in StonexStore code
+ * Allows to convert outer modifiers to Action/Module/Store specific modifiers
+ *
+ */
+export class ModifiersWorker {
 
   public static getModuleModifiers<MP> (modifiers: Array<Modifier<MP>>, storeInstance: Store<MP>): ModuleModifier[] {
     return modifiers.reduce((moduleModifiers: ModuleModifier[], modifier) => {